graph_tutorial/requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-sslserver==0.22
idna==2.9
jsonpickle==1.5.0
oauthlib==3.1.0
python-dateutil==2.8.1
python-http-client==3.3.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
sendgrid==6.5.0
six==1.14.0
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
urllib3==1.25.8